Change default dtype of attributes that are pandas series in Algorithm.

+++ algo_trading/algorithm.py
@@ -6,8 +6,8 @@ import pandas as pd
 class Algorithm(ABC):
     def __init__(self, capital: float = 10000):
         self.buying_power = capital
-        self.buys = pd.Series()
-        self.sells = pd.Series()
+        self.buys = pd.Series(dtype=int)
+        self.sells = pd.Series(dtype=int)
 
     @abstractmethod
     def setup(self):
